delete image with product

+++ controllers/admin.js
@@ -3,6 +3,7 @@ const mongoose = require("mongoose");
 const { validationResult } = require("express-validator");
 
 const Product = require("../models/product");
+const fileHelper = require("../util/file");
 
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
@@ -145,6 +146,7 @@ exports.postEditProduct = (req, res, next) => {
       product.description = updatedDesc;
       if (image) {
         product.imageUrl = image.path;
+        fileHelper.deleteFile(product.imageUrl);
       }
       return product.save().then((result) => {
         console.log("UPDATED PRODUCT!");
@@ -180,7 +182,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error("Product not found."));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log("DESTROYED PRODUCT");
       res.redirect("/admin/products");
